rename the package and remove the trackline in the cover
