+++ app/src/main/assets/litepal.xml
@@ -32,7 +32,7 @@
     	</list>
     -->
     <list>
-        <mapping class="com.sample.andremion.musicplayer.model.Song" />
+        <mapping class="team.fzo.puppas.mini_player.model.Song" />
     </list>
 
     <!--
