--- app/src/main/java/com/sample/andremion/musicplayer/view/InsetsPercentRelativeLayout.java
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2016. André Mion
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.sample.andremion.musicplayer.view;
-
-import android.content.Context;
-import android.support.percent.PercentRelativeLayout;
-import android.support.v4.view.ViewCompat;
-import android.support.v4.view.WindowInsetsCompat;
-import android.util.AttributeSet;
-import android.view.View;
-
-public class InsetsPercentRelativeLayout extends PercentRelativeLayout {
-
-    public InsetsPercentRelativeLayout(Context context) {
-        this(context, null, 0);
-    }
-
-    public InsetsPercentRelativeLayout(Context context, AttributeSet attrs) {
-        this(context, attrs, 0);
-    }
-
-    public InsetsPercentRelativeLayout(Context context, AttributeSet attrs, int defStyle) {
-        super(context, attrs, defStyle);
-        ViewCompat.setOnApplyWindowInsetsListener(this, new android.support.v4.view.OnApplyWindowInsetsListener() {
-            @Override
-            public WindowInsetsCompat onApplyWindowInsets(View v, WindowInsetsCompat insets) {
-                setWindowInsets(insets);
-                return insets.consumeSystemWindowInsets();
-            }
-        });
-    }
-
-    private void setWindowInsets(WindowInsetsCompat insets) {
-        // Now dispatch them to our children
-        for (int i = 0, z = getChildCount(); i < z; i++) {
-            final View child = getChildAt(i);
-            insets = ViewCompat.dispatchApplyWindowInsets(child, insets);
-            if (insets.isConsumed()) {
-                break;
-            }
-        }
-    }
-
-}
