+++ app/src/main/res/layout/activity_music_list.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.sample.andremion.musicplayer.view.InsetsPercentRelativeLayout
+<team.fzo.puppas.mini_player.view.InsetsPercentRelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -53,7 +53,7 @@
                 android:textAppearance="@style/TextAppearance.AppCompat.Caption"
                 android:transitionName="@string/transition_name_time" />
 
-            <com.sample.andremion.musicplayer.view.ProgressView
+            <team.fzo.puppas.mini_player.view.ProgressView
                 android:id="@+id/progress"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
@@ -142,4 +142,4 @@
         android:transitionName="@string/transition_name_fab"
         app:fabSize="mini"/>
 
-</com.sample.andremion.musicplayer.view.InsetsPercentRelativeLayout>
+</team.fzo.puppas.mini_player.view.InsetsPercentRelativeLayout>
