+++ app/src/main/res/transition/detail_shared_element_enter_transition.xml
@@ -4,7 +4,7 @@
     android:interpolator="@android:interpolator/accelerate_quad">
 
     <!--Progress-->
-    <transition class="com.sample.andremion.musicplayer.transition.ProgressViewTransition" />
+    <transition class="team.fzo.puppas.mini_player.transition.ProgressViewTransition" />
     <!--Cover-->
     <transition class="com.andremion.music.MusicCoverViewTransition" />
     <!--Others-->
